+++ lib/lz4.ts
@@ -1,0 +1,7 @@
+import { LZ4MaximumBlockSize, LZ4Dictionary, LZ4CompressionOptions } from "./lz4/options.js";
+import { LZ4Compressor } from "./lz4/compressor.js";
+
+export {
+    LZ4MaximumBlockSize, LZ4Dictionary, LZ4CompressionOptions,
+    LZ4Compressor
+};